Update actions/checkout to v4 with additional options
This commit updates the GitHub Actions workflow to use version 4 of the `actions/checkout` action. It adds parameters for `fetch-depth`, `ref`, and `fetch-tags`, enabling a full repository history fetch, specifying the branch to check out, and ensuring that tags are included during the checkout process.

Update .gitignore for Visual Studio workspace settings

Added entry to ignore the /.vs directory and included a comment indicating the file was created by Microsoft Visual Studio.